+++ lib/src/main/java/org/eclipse/tsp/java/client/api/timegraph/dto/GetTimeGraphTreeRequestDto.java
@@ -1,6 +1,5 @@
 package org.eclipse.tsp.java.client.api.timegraph.dto;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.eclipse.tsp.java.client.shared.query.QueryInterval;
@@ -16,7 +15,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class GetTimeGraphTreeRequestDto {
 	@JsonProperty("requested_times")
-	private List<BigInteger> requestedTimes;
+	private List<Long> requestedTimes;
 
 	@JsonProperty("requested_timerange")
 	private QueryInterval requestedTimerange;
